create mapper by userEntity
