+++ src/main/java/Estructure/domain/entities/UserEntity.java
@@ -6,7 +6,11 @@ import javax.management.relation.Role;
 import java.util.List;
 
 @Entity
-@Data
+@Setter
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
 public class UserEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
