+++ src/main/java/Estructure/infraestructure/dtos/request/GroupRequestDTO.java
@@ -1,10 +1,9 @@
-package Estructure.domain.dtos.response;
+package Estructure.infraestructure.dtos.request;
 
 import lombok.Data;
 
 @Data
-public class GroupResponseDTO {
-    private Long id;
+public class GroupRequestDTO {
     private Long projectId;
     private Long userId;
     private String groupName;
