+++ src/main/java/riwi/project/Application.java
@@ -1,7 +1,12 @@
 package riwi.project;
 
+import Estructure.infraestructure.dtos.request.UserRequestDTO;
+import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+
+import static Estructure.utils.Enums.Role.ADMIN;
 
 @SpringBootApplication
 public class Application {
@@ -9,5 +14,4 @@ public class Application {
 	public static void main(String[] args) {
 		SpringApplication.run(Application.class, args);
 	}
-
 }
